+++ src/main/java/org/jboss/forge/arquillian/api/algeron/AlgeronProviderFacet.java
@@ -2,7 +2,7 @@ package org.jboss.forge.arquillian.api.algeron;
 
 
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
-import org.jboss.forge.arquillian.container.model.ContractProviderLibrary;
+import org.jboss.forge.arquillian.model.algeron.ContractProviderLibrary;
 
 public class AlgeronProviderFacet extends AlgeronSetupFacet {
