+++ src/main/java/org/jboss/forge/arquillian/command/algeron/consumer/AlgeronCreateContractFragmentCommand.java
@@ -22,7 +22,7 @@ import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
 import org.jboss.forge.arquillian.api.algeron.AlgeronSetupFacet;
-import org.jboss.forge.arquillian.container.model.ContractConsumerLibrary;
+import org.jboss.forge.arquillian.model.algeron.ContractConsumerLibrary;
 import org.jboss.forge.arquillian.api.algeron.AlgeronConsumerFacet;
 import org.jboss.forge.roaster.model.JavaType;
 import org.jboss.forge.roaster.model.source.JavaClassSource;
