+++ src/main/java/org/jboss/forge/arquillian/command/algeron/retriever/AlgeronAddUrlRetrieverCommand.java
@@ -7,7 +7,7 @@ import org.jboss.forge.addon.ui.metadata.UICommandMetadata;
 import org.jboss.forge.addon.ui.metadata.WithAttributes;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.util.URLUIValidator;
+import org.jboss.forge.arquillian.validator.URLUIValidator;
 
 import javax.inject.Inject;
 import java.util.LinkedHashMap;
