+++ src/main/java/org/jboss/forge/arquillian/testframework/JUnitFacet.java
@@ -4,7 +4,7 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.testframework.junit;
+package org.jboss.forge.arquillian.testframework;
 
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
 import org.jboss.forge.arquillian.api.TestFrameworkFacet;
